Remove resource abstraction and wire peripherals
Remove the generated resources module and assign-resources crate.
Initialize and pass Peri peripherals directly from main to tasks. Add
bind_interrupts and a shared I2C mutex type; update ADC, I2C, PWM,
encoder, IMU, display, and other task signatures accordingly.

+++ src/task/autonomous_drive.rs
@@ -90,7 +90,7 @@ pub async fn autonomous_drive() {
 
                 // Random turn
                 // Use current time since boot as seed
-                let seed = Instant::now().as_micros() as u64;
+                let seed = Instant::now().as_micros();
                 let mut rng = WyRand::new_seed(seed);
 
                 info!("turning");
